requirements.txt
fastapi>=0.110.1
uvicorn>=0.22.0
pydantic>=2.8.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
supabase==2.15.0
PyJWT==2.10.1
greenlet==3.1.1
names-generator==0.2.0
shortuuid==1.0.13
substrate-interface==1.7.11
async-substrate-interface==1.2.1

